+++ src/trackers/core/sort/tracker.py
@@ -103,9 +103,13 @@ class SORTTracker(BaseTracker):
         self._reset_id_allocator()
 
     @property
-    @deprecated(target=None, deprecated_in="2.0", remove_in="3.0")
+    @deprecated(target=None, deprecated_in="2.5", remove_in="3.0")
     def trackers(self) -> list[SORTTracklet]:
-        """Deprecated: use tracks instead."""
+        """Deprecated alias for :attr:`tracks`.
+
+        .. deprecated:: 2.5
+            Use :attr:`tracks` instead. Will be removed in v3.0.
+        """
         return self.tracks
 
     def _get_associated_indices(